+++ stts/android/src/main/kotlin/com/llfbandit/stts/tts/model/TtsQueueMode.kt
@@ -1,0 +1,6 @@
+package com.llfbandit.stts.tts.model
+
+enum class TtsQueueMode {
+  Flush,
+  Add,
+}