+++ stts_platform_interface/lib/src/stt/model/model.dart
@@ -1,0 +1,4 @@
+export 'ios_audio_session.dart';
+export 'stt_recognition.dart';
+export 'stt_recognition_options.dart';
+export 'stt_state.dart';