+++ stts_platform_interface/lib/src/tts/model/model.dart
@@ -1,0 +1,4 @@
+export 'tts_options.dart';
+export 'tts_queue_mode.dart';
+export 'tts_state.dart';
+export 'tts_voice.dart';