+++ stts_platform_interface/lib/src/tts/model/tts_queue_mode.dart
@@ -1,0 +1,9 @@
+/// Text-to-Speech queue behaviour modes
+enum TtsQueueMode {
+  /// Cancels all current & next utterances
+  /// and replaces to the given one.
+  flush,
+
+  /// Adds an utterance to the current queue with the given text.
+  add,
+}