+++ stts_platform_interface/lib/stts_platform_interface.dart
@@ -1,8 +1,4 @@
-export 'src/stt/model/ios_audio_session.dart';
-export 'src/stt/model/stt_recognition.dart';
-export 'src/stt/model/stt_recognition_options.dart';
-export 'src/stt/model/stt_state.dart';
+export 'src/stt/model/model.dart';
 export 'src/stt/stt_platform_interface.dart';
-export 'src/tts/model/tts_state.dart';
-export 'src/tts/model/tts_voice.dart';
+export 'src/tts/model/model.dart';
 export 'src/tts/tts_platform_interface.dart';